remove float_to_string from core package
