+++ src/elfpy/utils/outputs.py
@@ -47,41 +47,6 @@ def annotate(axis_handle, text, major_offset, minor_offset, val):
     )
 
 
-def float_to_string(value, precision=3, min_digits=0, debug=False):
-    """
-    Format a float to a string with a given precision
-    this follows the significant figure behavior, irrepective of number size
-    """
-    # TODO: Include more specific error handling in the except statement
-    # pylint: disable=broad-except
-    if debug:
-        print(f"value: {value}, type: {type(value)}, precision: {precision}, min_digits: {min_digits}")
-    if np.isinf(value):
-        return "inf"
-    if np.isnan(value):
-        return "nan"
-    if value == 0:
-        return "0"
-    try:
-        digits = int(np.floor(np.log10(abs(value)))) + 1  #  calculate number of digits in value
-    except Exception as err:
-        if debug:
-            print(
-                f"Error in float_to_string: value={value}({type(value)}), precision={precision},"
-                f" min_digits={min_digits}, \n error={err}"
-            )
-        return str(value)
-    # decimals = np.clip(precision - digits, 0, precision)
-    decimals = min(max(precision - digits, min_digits), precision)  #  calculate desired decimals
-    if debug:
-        print(f"value: {value}, type: {type(value)} calculated digits: {digits}, decimals: {decimals}")
-    if abs(value) > 0.1:
-        string = f"{value:,.{decimals}f}"
-    else:  # add an additional sigfig if the value is really small
-        string = f"{value:0.{precision-1}e}"
-    return string
-
-
 def delete_log_file() -> None:
     """If the logger's handler if a file handler, delete the underlying file."""
     handler = logging.getLogger().handlers[0]
